Finish basic calculator functionalities and revamped UI
I finished implementing the basic arithmetic functionalities for the
calculator such as addition, subtraction, multiplication, and division.
I implemented helpers for the display, and I also finished the CE button
which resets the calculator. Lastly, I changed the UI to have a darker
background and removed the more advanced buttons of the calculator.